add custom useCharactersApi hook
